requirements.txt
flask
groq
python-dotenv
PyPDF2
